start updating mutation runner api, break many things

+++ posthog/clickhouse/cluster.py
@@ -393,31 +393,33 @@ class MutationNotFound(Exception):
 
 
 @dataclass
-class Mutation:
+class MutationWaiter:
     table: str
-    mutation_id: str
+    mutation_ids: Set[str]
 
     def is_done(self, client: Client) -> bool:
+        # TODO: this should maybe raise if the number of commands per mutation is incorrect?
         rows = client.execute(
             f"""
             SELECT
                 mutation_id,  -- ensure no rows are returned if the mutation we're looking for doesn't exist
-                countIf(is_done) = count()  -- multiple commands can be issued in a single mutation, consolidate all statuses into one value
+                countIf(is_done) = count() as all_commands_done -- multiple commands can be issued in a single mutation, consolidate all statuses into one value
             FROM system.mutations
-            WHERE database = %(database)s AND table = %(table)s AND mutation_id = %(mutation_id)s
+            WHERE database = %(database)s AND table = %(table)s AND mutation_id IN %(mutation_ids)s
             GROUP BY ALL
             """,
-            {"database": settings.CLICKHOUSE_DATABASE, "table": self.table, "mutation_id": self.mutation_id},
+            {"database": settings.CLICKHOUSE_DATABASE, "table": self.table, "mutation_ids": self.mutation_id},
         )
 
-        if len(rows) == 1:
-            [[mutation_id, is_done]] = rows
-            assert mutation_id == self.mutation_id
-            return is_done
-        elif len(rows) == 0:
-            raise MutationNotFound(f"could not find mutation matching {self!r}")
+        statuses = dict(rows)
+        assert len(rows) == len(statuses)
+
+        if missing_mutations := self.mutation_ids - statuses.keys():
+            raise MutationNotFound(f"could not find mutation(s): {missing_mutations!r}")
+        elif unexpected_mutations := statuses.keys() - self.mutation_ids:
+            raise ValueError(f"received unexpected mutation(s): {unexpected_mutations!r}")  # should never happen
         else:
-            raise ValueError(f"expected zero or one rows, found {len(rows)}")
+            return all(statuses.values())
 
     def wait(self, client: Client) -> None:
         while not self.is_done(client):
@@ -431,22 +433,48 @@ class MutationRunner(abc.ABC):
     settings: Mapping[str, Any] = field(default_factory=dict, kw_only=True)
 
     @abc.abstractmethod
-    def get_statement(self) -> str:
-        """Returns the full statement that can be used to enqueue the mutation."""
+    def get_commands(self) -> Set[str]:
+        """Returns the command that can be used to find the mutation in the ``system.mutations`` table."""
         raise NotImplementedError
 
     @abc.abstractmethod
-    def get_commands(self) -> Set[str]:
-        """Returns the command that can be used to find the mutation in the ``system.mutations`` table."""
+    def get_statement(self, commands: Set[str]) -> str:
+        """Returns the full statement that can be used to enqueue the mutation for the provided commands."""
         raise NotImplementedError
 
     def __post_init__(self) -> None:
         if invalid_keys := {key for key in self.parameters.keys() if key.startswith("__")}:
             raise ValueError(f"invalid parameter names: {invalid_keys!r} (keys cannot start with double underscore)")
 
-    def _find(self, client: Client) -> Mutation | None:
-        """Find the running mutation task, if one exists."""
-        commands = [*self.get_commands()]
+    def __call__(self, client: Client) -> MutationWaiter:
+        """
+        Ensure that all mutation commands are either running, or have previously run to completion. Returns an object
+        that can be used to check the status of the mutation and wait for its completion.
+        """
+        expected_commands = self.get_commands()
+        mutations = self.__find(client, expected_commands)
+
+        commands_to_enqueue = {command for command, mutation in mutations.items() if mutation is None}
+        if commands_to_enqueue:
+            client.execute(self.get_statement(commands_to_enqueue), self.parameters, settings=self.settings)
+
+        # find mutation ids for all commands
+        # mutations are not always immediately visible, so give it a bit of time to show up
+        start = time.time()
+        for _ in range(5):
+            mutations = self.__find(client, expected_commands)
+            if all(mutations.values()):
+                return MutationWaiter(self.table, set(mutations.values()))  # TODO
+
+        raise Exception(f"unable to find mutation after {time.time() - start:0.2f}s!")
+
+    def __find(self, client: Client, commands: Set[str]) -> Mapping[str, str | None]:
+        """
+        Find the mutation ID (if it exists) associated with each command provided.
+        """
+        # we match commands by position, so require a stable ordering - this is because this class is provided the
+        # parameterized generic command, while the record we match in the mutation log will have the parameters inlined
+        command_list = [*commands]
         mutations = client.execute(
             f"""
             SELECT mutation_id
@@ -457,7 +485,7 @@ class MutationRunner(abc.ABC):
                             arrayMap(
                                 command -> extract(command, '^\\s*(.*?)(?:,)?\\s*$'),  -- strip leading/trailing whitespace and optional trailing comma
                                 arraySlice(  -- drop "ALTER TABLE" preamble line
-                                    splitByChar('\n', formatQuery($__sql$ALTER TABLE {settings.CLICKHOUSE_DATABASE}.{self.table} {", ".join(commands)}$__sql$)),
+                                    splitByChar('\n', formatQuery($__sql$ALTER TABLE {settings.CLICKHOUSE_DATABASE}.{self.table} {", ".join(command_list)}$__sql$)),
                                     2
                                 )
                             ) as commands,
@@ -482,32 +510,8 @@ class MutationRunner(abc.ABC):
                 **self.parameters,
             },
         )
-        assert len(mutations) == len(commands)
-        command_mutations = {command: mutation for command, (mutation,) in zip(commands, mutations)}
-
-        if not any(command_mutations.values()):
-            return None
-        else:
-            # TODO: handle multiple results if a mutation was split across multiple runs, such as after a code change
-            # that adds another command to an existing mutation
-            [mutation_id] = set(command_mutations.values())
-            return Mutation(self.table, mutation_id)
-
-    def enqueue(self, client: Client) -> Mutation:
-        """Enqueue the mutation (or return the existing mutation if it is already running or has run.)"""
-        if task := self._find(client):
-            return task
-
-        client.execute(self.get_statement(), self.parameters, settings=self.settings)
-
-        # mutations are not always immediately visible, so give it a bit of time to show up
-        start = time.time()
-        for _ in range(5):
-            if task := self._find(client):
-                return task
-            time.sleep(1.0)
-
-        raise Exception(f"unable to find mutation after {time.time() - start:0.2f}s!")
+        assert len(mutations) == len(command_list)
+        return {command: mutation_id for command, (mutation_id,) in zip(command_list, mutations)}
 
     def run_on_shards(self, cluster: ClickhouseCluster, shards: Iterable[int] | None = None) -> None:
         """
@@ -515,24 +519,24 @@ class MutationRunner(abc.ABC):
         hosts within the affected shards.
         """
         if shards is not None:
-            shard_host_mutations = cluster.map_any_host_in_shards({shard: self.enqueue for shard in shards})
+            shard_host_mutation_waiters = cluster.map_any_host_in_shards({shard: self for shard in shards})
         else:
-            shard_host_mutations = cluster.map_one_host_per_shard(self.enqueue)
+            shard_host_mutation_waiters = cluster.map_one_host_per_shard(self)
 
         # XXX: need to convert the `shard_num` of type `int | None` to `int` to appease the type checker -- but nothing
         # should have actually been filtered out, since we're using the cluster shard functions for targeting
         shard_mutations = {
             host.shard_num: mutations
-            for host, mutations in shard_host_mutations.result().items()
+            for host, mutations in shard_host_mutation_waiters.result().items()
             if host.shard_num is not None
         }
-        assert len(shard_mutations) == len(shard_host_mutations)
+        assert len(shard_mutations) == len(shard_host_mutation_waiters)
 
         # during periods of elevated replication lag, it may take some time for mutations to become available on
         # the shards, so give them a little bit of breathing room with retries
         retry_policy = RetryPolicy(max_attempts=3, delay=10.0, exceptions=(MutationNotFound,))
         cluster.map_all_hosts_in_shards(
-            {shard_num: retry_policy(mutation.wait) for shard_num, mutation in shard_mutations.items()}
+            {shard_num: retry_policy(waiter) for shard_num, waiter in shard_mutations.items()}
         ).result()
 
 
@@ -540,19 +544,23 @@ class MutationRunner(abc.ABC):
 class AlterTableMutationRunner(MutationRunner):
     commands: Set[str]  # the part after ALTER TABLE prefix, i.e. UPDATE, DELETE, MATERIALIZE, etc.
 
-    def get_statement(self) -> str:
-        return f"ALTER TABLE {settings.CLICKHOUSE_DATABASE}.{self.table} " + ", ".join(self.commands)
-
     def get_commands(self) -> Set[str]:
         return self.commands
+
+    def get_statement(self, commands: Set[str]) -> str:
+        return f"ALTER TABLE {settings.CLICKHOUSE_DATABASE}.{self.table} " + ", ".join(commands)
 
 
 @dataclass
 class LightweightDeleteMutationRunner(MutationRunner):
     predicate: str
 
-    def get_statement(self) -> str:
-        return f"DELETE FROM {settings.CLICKHOUSE_DATABASE}.{self.table} WHERE {self.predicate}"
-
     def get_commands(self) -> Set[str]:
         return {f"UPDATE _row_exists = 0 WHERE {self.predicate}"}
+
+    def get_statement(self, commands: Set[str]) -> str:
+        # XXX: lightweight deletes should only be called with the same command represented by the predicate
+        if commands != self.get_commands():
+            raise ValueError(f"unexpected commands: {commands!r}")
+
+        return f"DELETE FROM {settings.CLICKHOUSE_DATABASE}.{self.table} WHERE {self.predicate}"
